Data Engineering - eCommerce/spark/requirements.txt
pyspark==3.2.0
mysql-connector-python==8.0.28